Send plugin status to OPA
There are new plugin status API's available on the OPA plugin manager
which can allow a plugin to affect the health state for the OPA
server. This sets up some basic status lifecycle points for the
plugin.

Note that due to the way the Golang API works for gRPC there is a race
between switching the state to OK and starting to handle incomming
requests. There doesn't appear to be a way to check the state of the
listening server after we make the blocking call to
`p.server.Serve(l)`. This doesn't seem like it would cause any issues
however envoy/istio can handle errors on the grpc interface. We are
mostly concerned if it goes into an error or not-ready state.

Signed-off-by: Patrick East <east.patrick@gmail.com>